Reduce dependency on Node.js core internals
This commit reduces the ways in which `minizlib` depends on
Node.js core internals. It does not move this module to purely
public API, but this should allow for more flexibility in
modifying Node.js core.

In particular, the semi-public `_processChunk` method can be used
instead of imitating the way Node.js core performs writes on the
native handle.

This patch also removes access to `process.binding()`; instead,
the Node.js core zlib wrapper serves as the “handle” for minizlib.