make brotli work, remove unnecessary error checking
The classes defined in this library used to do some extensive checking
to ensure that every parameter is within the allowed range.

This was important, because we talked directly to the C++ binding layer
(via a not-officially supported API).

Some time ago, this module started using Node's JS zlib classes as our
'_handle', meaning that this code is doing double-work, since the JS
layer zlib classes do the same type and range checking.

The downside is that the error messages change between Node.js versions,
so we'll just have to accept that 'throws' is a safe enough test
description.

The instigator for this error handling simplification was splitting out
the Brotli classes properly, since all of their parameters would have to
be tracked and checked, and Brotli has a lot of those.

Better to just let node do it.